Step5: 추상 리스너 생성
리스너에 구현은 하지 않았습니다
