--- practice-order/src/main/java/com/sparta/practiceorder/dto/OrderResponse.java
@@ -1,19 +1,0 @@
-package com.sparta.practiceorder.dto;
-
-import com.sparta.practiceorder.entity.Order;
-import java.util.UUID;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class OrderResponse {
-
-    UUID id;
-
-    public static OrderResponse from(Order savedOrder) {
-        return new OrderResponse(
-            savedOrder.getId()
-        );
-    }
-}
