--- practice-order/src/main/java/com/sparta/practiceorder/entity/Product.java
@@ -1,29 +1,0 @@
-package com.sparta.practiceorder.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import java.util.UUID;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class Product {
-    @Id
-    private UUID id;
-
-    private String name;
-    private int stock; // 재고
-
-    public void decreaseStock(int quantity) {
-        if (this.stock < quantity) {
-            throw new IllegalArgumentException("재고가 부족합니다");
-        }
-        this.stock -= quantity;
-    }
-}
