--- practice-order/src/main/java/com/sparta/practiceorder/entity/enums/PaymentStatus.java
@@ -1,7 +1,0 @@
-package com.sparta.practiceorder.entity.enums;
-
-public enum PaymentStatus {
-    PENDING,
-    COMPLETED,
-    FAILED
-}
