--- practice-order/src/main/java/com/sparta/practiceorder/enums/OrderStatus.java
@@ -1,6 +1,0 @@
-package com.sparta.practiceorder.enums;
-
-public enum OrderStatus {
-    PENDING,
-
-}
