+++ practice-order/src/main/java/com/sparta/practiceorder/order/domain/enums/OrderStatus.java
@@ -1,0 +1,6 @@
+package com.sparta.practiceorder.order.domain.enums;
+
+public enum OrderStatus {
+    PENDING,
+
+}
