+++ practice-order/src/main/java/com/sparta/practiceorder/order/dto/OrderResponse.java
@@ -1,0 +1,19 @@
+package com.sparta.practiceorder.order.dto;
+
+import com.sparta.practiceorder.order.domain.Order;
+import java.util.UUID;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class OrderResponse {
+
+    UUID id;
+
+    public static OrderResponse from(Order savedOrder) {
+        return new OrderResponse(
+            savedOrder.getId()
+        );
+    }
+}
