+++ practice-order/src/main/java/com/sparta/practiceorder/payment/domain/Payment.java
@@ -1,0 +1,50 @@
+package com.sparta.practiceorder.payment.domain;
+
+import com.sparta.practiceorder.payment.domain.enums.PaymentStatus;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+import java.util.UUID;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Payment {
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
+
+    private UUID orderId;
+
+    @Enumerated(EnumType.STRING)
+    private PaymentStatus status; // PENDING, COMPLETED, FAILED
+
+    private BigDecimal amount;
+
+    private LocalDateTime createdAt;
+
+    public static Payment create(UUID orderId, BigDecimal amount) {
+        Payment payment = new Payment();
+        payment.orderId = orderId;
+        payment.amount = amount;
+        payment.status = PaymentStatus.PENDING;
+        payment.createdAt = LocalDateTime.now();
+        return payment;
+    }
+
+    public void complete() {
+        this.status = PaymentStatus.COMPLETED;
+    }
+
+    public void fail() {
+        this.status = PaymentStatus.FAILED;
+    }
+}
