+++ practice-order/src/main/java/com/sparta/practiceorder/payment/domain/enums/PaymentStatus.java
@@ -1,0 +1,7 @@
+package com.sparta.practiceorder.payment.domain.enums;
+
+public enum PaymentStatus {
+    PENDING,
+    COMPLETED,
+    FAILED
+}
