+++ practice-order/src/main/java/com/sparta/practiceorder/payment/repository/PaymentRepository.java
@@ -1,0 +1,11 @@
+package com.sparta.practiceorder.payment.repository;
+
+import com.sparta.practiceorder.payment.domain.Payment;
+import java.util.List;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PaymentRepository extends JpaRepository<Payment, UUID> {
+
+    List<Payment> findByOrderId(UUID orderId);
+}
