--- practice-order/src/main/java/com/sparta/practiceorder/repository/OrderRepository.java
@@ -1,9 +1,0 @@
-package com.sparta.practiceorder.repository;
-
-import com.sparta.practiceorder.entity.Order;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<Order, UUID> {
-
-}
