--- practice-order/src/main/java/com/sparta/practiceorder/repository/PaymentRepository.java
@@ -1,11 +1,0 @@
-package com.sparta.practiceorder.repository;
-
-import com.sparta.practiceorder.entity.Payment;
-import java.util.List;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PaymentRepository extends JpaRepository<Payment, UUID> {
-
-    List<Payment> findByOrderId(UUID orderId);
-}
