--- practice-order/src/main/java/com/sparta/practiceorder/repository/ProductRepository.java
@@ -1,9 +1,0 @@
-package com.sparta.practiceorder.repository;
-
-import com.sparta.practiceorder.entity.Product;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, UUID> {
-
-}
